Move elasticjob-error-handler-general to general package (#1522)
* Move elasticjob-error-handler-general to general package

* Rename package
